pyproject.toml
[tool.poetry]
name = "thz-tds-processing"
version = "0.1.0"
description = "THz-TDS Time Domain Signal Processing and FFT Analysis Pipeline"
authors = ["Alperen Sari"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
numpy = "^1.24"
pandas = "^2.0"
scipy = "^1.10"
typer = "^0.20.0"
tqdm = "^4.65"
pyyaml = "^6.0"
pydantic = "^2.0"
matplotlib = "^3.7"
requests = "^2.32.5"
click = "^8.3.1"
rich = "^14.2.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
tds-process = "src.main:app"